fix printf compatibility bug (this is why you don't code at home on linux)

+++ libft/src/ft_pr_convert.c
@@ -6,7 +6,7 @@
 /*   By: ygaude <ygaude@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/05 20:05:45 by ygaude            #+#    #+#             */
-/*   Updated: 2018/03/09 20:42:33 by ygaude           ###   ########.fr       */
+/*   Updated: 2018/03/10 17:56:52 by sbrochar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ t_str	ft_unic(wchar_t unicstr[], int justonce, int prec, char **str)
 	{
 		c = *unicstr;
 		i = (c >= 0x80) + (c >= 0x800) + (c >= 0x10000) + (c >= 0x110000);
-		if ((unsigned int)i + 1 > MB_CUR_MAX)
+		if ((unsigned int)i + 1 > (unsigned int)MB_CUR_MAX)
 		{
 			ft_strdel(&(res.str));
 			return (res);
